backend/requirements.txt
Flask
python-dotenv
Requests
requests
numpy
argon2-cffi
pandas
matplotlib
seaborn
flask_sqlalchemy
Flask_CORS
SQLAlchemy
pyodbc
fastapi
uvicorn[standard]
pydantic
azure-identity
geoalchemy2
shapely
